moved app, security and placeholder middleware functions
