+++ security/getUnauthorizedResponse.js
